// rmt-ui/src/common/interfaces/IBuTreeMapping.ts
export interface IBUTreeMapping {
  bu: string;
  bu_id: string;
  bu_leader_mid?: string;
  offering: string;
  offering_id: string;
  offering_leader_mid?: string;
  solution: string;
  solution_leader_mid?: string;
  solution_id: string;
  bu_efficiency_leader_mid?: string;
  isactive?: boolean;
  createdat?: Date;
  modifiedat?: Date;
  createdby?: string;
  modifiedby?: string;
}